Organized voting page files.
