+++ together-wherever/src/app/constants/vote.ts
@@ -1,0 +1,4 @@
+export const VOTING_CONFIRMATION_DIALOG = {
+    topic: "Complete Vote",
+    desc: "Please confirm that you want to complete your vote."
+};
